fix(tests): Moved to correct namespace
