+++ Core/Tests/ExceptionTests.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using NUnit.Framework;
 using Speckle.Core.Models.Extensions;
-namespace TestsUnit
+namespace Tests
 {
   [TestFixture]
   public class ExceptionTests
