+++ Core/Tests/Helpers.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using NUnit.Framework;
 
-namespace TestsUnit
+namespace Tests
 {
   [TestFixture]
   public class Helpers
